Better table structure.
